feat(home): completed homepage hero section
