+++ src/app/globals.css
@@ -27,6 +27,22 @@
     animation: gradient-shift 3s ease infinite;
   }
 
+  .animate-workflow-rise {
+    animation: workflow-rise 0.7s ease both;
+  }
+
+  .animate-workflow-scan {
+    animation: workflow-scan 4.5s ease-in-out infinite;
+  }
+
+  .animate-workflow-pulse {
+    animation: workflow-pulse 2.4s ease-in-out infinite;
+  }
+
+  .animate-workflow-border {
+    animation: workflow-border 5s linear infinite;
+  }
+
   @keyframes gradient-shift {
     0% {
       background-position: 0% 50%;
@@ -36,6 +52,64 @@
     }
     100% {
       background-position: 0% 50%;
+    }
+  }
+
+  @keyframes workflow-rise {
+    from {
+      opacity: 0;
+      transform: translateY(10px);
+    }
+    to {
+      opacity: 1;
+      transform: translateY(0);
+    }
+  }
+
+  @keyframes workflow-scan {
+    0%,
+    100% {
+      transform: translateX(-120%);
+      opacity: 0;
+    }
+    18%,
+    52% {
+      opacity: 0.8;
+    }
+    70% {
+      transform: translateX(120%);
+      opacity: 0;
+    }
+  }
+
+  @keyframes workflow-pulse {
+    0%,
+    100% {
+      box-shadow: 0 0 0 0 rgba(34, 211, 238, 0.35);
+      opacity: 0.85;
+    }
+    50% {
+      box-shadow: 0 0 0 8px rgba(34, 211, 238, 0);
+      opacity: 1;
+    }
+  }
+
+  @keyframes workflow-border {
+    0% {
+      transform: translateX(-70%) translateY(-70%) rotate(0deg);
+    }
+    100% {
+      transform: translateX(-70%) translateY(-70%) rotate(360deg);
+    }
+  }
+
+  @media (prefers-reduced-motion: reduce) {
+    .animate-gradient,
+    .animate-workflow-rise,
+    .animate-workflow-scan,
+    .animate-workflow-pulse,
+    .animate-workflow-border {
+      animation: none;
     }
   }
 }
